Rename packages to enable use of multiple Lyo clients in the same project
