+++ lyo-client-v4/src/main/java/org/eclipse/lyo/client/v4/OslcClientFactory.java
@@ -1,0 +1,14 @@
+package org.eclipse.lyo.client.v4;
+
+public class OslcClientFactory {
+    
+    static public OslcOAuthClientBuilder oslcOAuthClientBuilder() {
+        return new OslcOAuthClientBuilder();
+    }
+    
+    static public OslcClientBuilder oslcClientBuilder() {
+        return new OslcClientBuilder();
+    }
+
+
+}
