+++ lyo-client-v4/src/main/resources/oslc4jclient.properties
