+++ lyo-client-v4/src/test/resources/blogQuery.rdf
