+++ lyo-client-v4/src/test/resources/emptyQuery.rdf
