+++ lyo-client-v4/src/test/resources/noParamQuery.rdf
