+++ lyo-client-v4/src/test/resources/queryResponse.rdf
