--- lyo-client/src/main/java/org/eclipse/lyo/client/OslcClientFactory.java
@@ -1,14 +1,0 @@
-package org.eclipse.lyo.client;
-
-public class OslcClientFactory {
-    
-    static public OslcOAuthClientBuilder oslcOAuthClientBuilder() {
-        return new OslcOAuthClientBuilder();
-    }
-    
-    static public OslcClientBuilder oslcClientBuilder() {
-        return new OslcClientBuilder();
-    }
-
-
-}
