--- lyo-client/src/main/java/org/eclipse/lyo/client/UnderlyingHttpClient.java
@@ -1,12 +1,0 @@
-package org.eclipse.lyo.client;
-
-import javax.ws.rs.client.Client;
-import org.apache.http.client.HttpClient;
-
-/**
- * Returns the Apache HTTP client underlying the JAX-RS client.
- * @return the http client
- */
-public interface UnderlyingHttpClient {
-    public HttpClient get(Client client);
-}
